RevolvingCounter use Long.MAX_VALUE This close #1172

+++ restcomm/restcomm.mgcp/src/main/java/org/mobicents/servlet/restcomm/mgcp/MediaGateway.java
@@ -228,9 +228,9 @@ public final class MediaGateway extends UntypedActor implements JainMgcpListener
         domain = new StringBuilder().append(remoteIp.getHostAddress()).append(":").append(remotePort).toString();
         notificationListeners.clear();
         responseListeners.clear();
-        requestIdPool = new RevolvingCounter(1, Integer.MAX_VALUE);
-        sessionIdPool = new RevolvingCounter(1, Integer.MAX_VALUE);
-        transactionIdPool = new RevolvingCounter(1, Integer.MAX_VALUE);
+        requestIdPool = new RevolvingCounter(1, Long.MAX_VALUE);
+        sessionIdPool = new RevolvingCounter(1, Long.MAX_VALUE);
+        transactionIdPool = new RevolvingCounter(1, Long.MAX_VALUE);
     }
 
     @Override
